+++ typings/azure-storage/azure-storage.d.ts
@@ -5941,7 +5941,7 @@ declare module azurestorage {
         * @param {LocationMode}       [options.locationMode]                      Specifies the location mode used to decide which location the request should be sent to.
         *                                                                         Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.maxResults]                        Specifies the maximum number of shares to return per call to Azure storage.
-        * @param {string}             [options.include]                           Include this parameter to specify that the share's metadata be returned as part of the response body. (allowed values: '', 'metadata')
+        * @param {string}             [options.include]                           Include this parameter to specify that the share's metadata be returned as part of the response body. (allowed values: '', 'metadata', 'snapshots' or any combination of them)
         * @param {int}                [options.timeoutIntervalInMs]               The server timeout interval, in milliseconds, to use for the request.
         * @param {int}                [options.maximumExecutionTimeInMs]          The maximum execution time, in milliseconds, across all potential retries, to use when making this request.
         *                                                                         The maximum execution time interval begins at the time that the client begins building the request. The maximum
@@ -5954,7 +5954,7 @@ declare module azurestorage {
         *                                                                         `entries`  gives a list of shares and the `continuationToken` is used for the next listing operation.
         *                                                                         `response` will contain information related to this operation.
         */
-        listSharesSegmented(currentToken: common.ContinuationToken, options: FileService.ListRequestOptions, callback: ErrorOrResult<FileService.ListSharesResult>): void;
+        listSharesSegmented(currentToken: common.ContinuationToken, options: FileService.ListShareRequestOptions, callback: ErrorOrResult<FileService.ListSharesResult>): void;
         listSharesSegmented(currentToken: common.ContinuationToken, callback: ErrorOrResult<FileService.ListSharesResult>): void;
 
         /**
@@ -5968,7 +5968,7 @@ declare module azurestorage {
         *                                                                         Please see StorageUtilities.LocationMode for the possible values.
         * @param {string}             [options.prefix]                            Filters the results to return only shares whose name begins with the specified prefix.
         * @param {int}                [options.maxResults]                        Specifies the maximum number of shares to return per call to Azure storage.
-        * @param {string}             [options.include]                           Include this parameter to specify that the share's metadata be returned as part of the response body. (allowed values: '', 'metadata')
+        * @param {string}             [options.include]                           Include this parameter to specify that the share's metadata be returned as part of the response body. (allowed values: '', 'metadata', 'snapshots' or any combination of them)
         * @param {int}                [options.timeoutIntervalInMs]               The server timeout interval, in milliseconds, to use for the request.
         * @param {int}                [options.maximumExecutionTimeInMs]          The maximum execution time, in milliseconds, across all potential retries, to use when making this request.
         *                                                                         The maximum execution time interval begins at the time that the client begins building the request. The maximum
@@ -5981,7 +5981,7 @@ declare module azurestorage {
         *                                                                         `entries`  gives a list of shares and the `continuationToken` is used for the next listing operation.
         *                                                                         `response` will contain information related to this operation.
         */
-        listSharesSegmentedWithPrefix(prefix: string, currentToken: common.ContinuationToken, options: FileService.ListRequestOptions, callback: ErrorOrResult<FileService.ListSharesResult>): void;
+        listSharesSegmentedWithPrefix(prefix: string, currentToken: common.ContinuationToken, options: FileService.ListShareRequestOptions, callback: ErrorOrResult<FileService.ListSharesResult>): void;
         listSharesSegmentedWithPrefix(prefix: string, currentToken: common.ContinuationToken, callback: ErrorOrResult<FileService.ListSharesResult>): void;
 
         /**
@@ -5990,6 +5990,7 @@ declare module azurestorage {
         * @this {FileService}
         * @param {string}             share                                   The share name.
         * @param {Object}             [options]                               The request options.
+        * @param {string}             [options.shareSnapshotId]               The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]                  Specifies the location mode used to decide which location the request should be sent to.
         *                                                                     Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]           The server timeout interval, in milliseconds, to use for the request.
@@ -6004,7 +6005,7 @@ declare module azurestorage {
         *                                                                     be true if the share exists, or false if the share does not exist.
         *                                                                     `response` will contain information related to this operation.
         */
-        doesShareExist(share: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
+        doesShareExist(share: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
         doesShareExist(share: string, callback: ErrorOrResult<FileService.ShareResult>): void;
 
         /**
@@ -6032,6 +6033,9 @@ declare module azurestorage {
         */
         createShare(share: string, options: FileService.CreateShareRequestOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
         createShare(share: string, callback: ErrorOrResult<FileService.ShareResult>): void;
+
+        createShareSnapshot(share: string, options: common.RequestOptions, callback: ErrorOrResult<string>): void;        
+        createShareSnapshot(share: string, callback: ErrorOrResult<string>): void;
 
         /**
         * Creates a new share under the specified account if the share does not exists.
@@ -6074,6 +6078,7 @@ declare module azurestorage {
         * @this {FileService}
         * @param {string}             share                               The share name.
         * @param {Object}             [options]                           The request options.
+        * @param {string}             [options.shareSnapshotId]           The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]              Specifies the location mode used to decide which location the request should be sent to.
         *                                                                 Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]       The server timeout interval, in milliseconds, to use for the request.
@@ -6088,7 +6093,7 @@ declare module azurestorage {
         *                                                                 information for the share.
         *                                                                 `response` will contain information related to this operation.
         */
-        getShareProperties(share: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
+        getShareProperties(share: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
         getShareProperties(share: string, callback: ErrorOrResult<FileService.ShareResult>): void;
 
         /**
@@ -6141,6 +6146,7 @@ declare module azurestorage {
         * @this {FileService}
         * @param {string}             share                                     The share name.
         * @param {Object}             [options]                                 The request options.
+        * @param {string}             [options.shareSnapshotId]                 The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]                    Specifies the location mode used to decide which location the request should be sent to.
         *                                                                       Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]             The server timeout interval, in milliseconds, to use for the request.
@@ -6155,7 +6161,7 @@ declare module azurestorage {
         *                                                                       information for the share.
         *                                                                       `response` will contain information related to this operation.
         */
-        getShareMetadata(share: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
+        getShareMetadata(share: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.ShareResult>): void;
         getShareMetadata(share: string, callback: ErrorOrResult<FileService.ShareResult>): void;
 
         /**
@@ -6190,6 +6196,8 @@ declare module azurestorage {
         * @this {FileService}
         * @param {string}             share                               The share name.
         * @param {Object}             [options]                           The request options.
+        * @param {string}             [options.deleteSnapshots]           The snapshot delete option. See azure.FileUtilities.ShareSnapshotDeleteOptions.*. 
+        * @param {string}             [options.shareSnapshotId]           The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]              Specifies the location mode used to decide which location the request should be sent to.
         *                                                                 Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]       The server timeout interval, in milliseconds, to use for the request.
@@ -6236,6 +6244,8 @@ declare module azurestorage {
         * @this {FileService}
         * @param {string}             share                               The share name.
         * @param {Object}             [options]                           The request options.
+        * @param {string}             [options.deleteSnapshots]           The snapshot delete option. See azure.FileUtilities.ShareSnapshotDeleteOptions.*. 
+        * @param {string}             [options.shareSnapshotId]           The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]              Specifies the location mode used to decide which location the request should be sent to.
         *                                                                 Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]       The server timeout interval, in milliseconds, to use for the request.
@@ -6249,7 +6259,7 @@ declare module azurestorage {
         *                                                                 if an error occurs; otherwise
         *                                                                 `response` will contain information related to this operation.
         */
-        deleteShare(share: string, options: common.RequestOptions, callback: ErrorOrResponse): void;
+        deleteShare(share: string, options: FileService.DeleteShareOptions, callback: ErrorOrResponse): void;
         deleteShare(share: string, callback: ErrorOrResponse): void;
 
         /**
@@ -6274,7 +6284,7 @@ declare module azurestorage {
         *                                                                 did not exist.
         *                                                                 `response` will contain information related to this operation.
         */
-        deleteShareIfExists(share: string, options: common.RequestOptions, callback: ErrorOrResult<boolean>): void;
+        deleteShareIfExists(share: string, options: FileService.DeleteShareOptions, callback: ErrorOrResult<boolean>): void;
         deleteShareIfExists(share: string, callback: ErrorOrResult<boolean>): void;
 
         /**
@@ -6284,6 +6294,7 @@ declare module azurestorage {
         * @param {string}             share                                   The share name.
         * @param {string}             directory                               The directory name. Use '' to refer to the base directory.
         * @param {Object}             [options]                               The request options.
+        * @param {string}             [options.shareSnapshotId]               The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]                  Specifies the location mode used to decide which location the request should be sent to.
         *                                                                     Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]           The server timeout interval, in milliseconds, to use for the request.
@@ -6298,7 +6309,7 @@ declare module azurestorage {
         *                                                                     be true if the directory exists, or false if the directory does not exist.
         *                                                                     `response` will contain information related to this operation.
         */
-        doesDirectoryExist(share: string, directory: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.DirectoryResult>): void;
+        doesDirectoryExist(share: string, directory: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.DirectoryResult>): void;
         doesDirectoryExist(share: string, directory: string, callback: ErrorOrResult<FileService.DirectoryResult>): void;
 
         /**
@@ -6367,6 +6378,7 @@ declare module azurestorage {
         * @param {string}             share                               The share name.
         * @param {string}             directory                           The directory name. Use '' to refer to the base directory.
         * @param {Object}             [options]                           The request options.
+        * @param {string}             [options.shareSnapshotId]           The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]              Specifies the location mode used to decide which location the request should be sent to.
         *                                                                 Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]       The server timeout interval, in milliseconds, to use for the request.
@@ -6381,7 +6393,7 @@ declare module azurestorage {
         *                                                                 information for the directory.
         *                                                                 `response` will contain information related to this operation.
         */
-        getDirectoryProperties(share: string, directory: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.DirectoryResult>): void;
+        getDirectoryProperties(share: string, directory: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.DirectoryResult>): void;
         getDirectoryProperties(share: string, directory: string, callback: ErrorOrResult<FileService.DirectoryResult>): void;
 
         /**
@@ -6440,6 +6452,7 @@ declare module azurestorage {
         * @param {string}             directory                         The directory name. Use '' to refer to the base directory.
         * @param {Object}             currentToken                      A continuation token returned by a previous listing operation. Please use 'null' or 'undefined' if this is the first operation.
         * @param {Object}             [options]                         The request options.
+        * @param {string}             [options.shareSnapshotId]         The snapshot identifier of the share.
         * @param {int}                [options.maxResults]              Specifies the maximum number of files to return per call to Azure ServiceClient. This does NOT affect list size returned by this function. (maximum: 5000)
         * @param {LocationMode}       [options.locationMode]            Specifies the location mode used to decide which location the request should be sent to.
         *                                                               Please see StorageUtilities.LocationMode for the possible values.
@@ -6468,6 +6481,7 @@ declare module azurestorage {
         * @param {string}             prefix                            The prefix of the directory/files name.
         * @param {Object}             currentToken                      A continuation token returned by a previous listing operation. Please use 'null' or 'undefined' if this is the first operation.
         * @param {Object}             [options]                         The request options.
+        * @param {string}             [options.shareSnapshotId]         The snapshot identifier of the share.
         * @param {int}                [options.maxResults]              Specifies the maximum number of files to return per call to Azure ServiceClient. This does NOT affect list size returned by this function. (maximum: 5000)
         * @param {LocationMode}       [options.locationMode]            Specifies the location mode used to decide which location the request should be sent to.
         *                                                               Please see StorageUtilities.LocationMode for the possible values.
@@ -6492,6 +6506,8 @@ declare module azurestorage {
         * @this {FileService}
         * @param {string}             share                                       The share name.
         * @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
+        * @param {object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {int}                [options.timeoutIntervalInMs]               The server timeout interval, in milliseconds, to use for the request.
         * @param {int}                [options.maximumExecutionTimeInMs]          The maximum execution time, in milliseconds, across all potential retries, to use when making this request.
         *                                                                         The maximum execution time interval begins at the time that the client begins building the request. The maximum
@@ -6504,7 +6520,7 @@ declare module azurestorage {
         *                                                                         information about the file.
         *                                                                         `response` will contain information related to this operation.
         */
-        getDirectoryMetadata(share: string, directory: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.DirectoryResult>): void;
+        getDirectoryMetadata(share: string, directory: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.DirectoryResult>): void;
         getDirectoryMetadata(share: string, directory: string, callback: ErrorOrResult<FileService.DirectoryResult>): void;
 
         /**
@@ -6583,13 +6599,14 @@ declare module azurestorage {
         * @param {string}                   [file]                   The file name. File names may not start or end with the delimiter '/'.
         * @param {string}                   [sasToken]               The Shared Access Signature token.
         * @param {boolean}                  [primary]                A boolean representing whether to use the primary or the secondary endpoint.
+        * @param {boolean}                  [shareSnapshotId]        The snapshot identifier of the share.
         * @return {string}                                           The formatted URL string.
         * @example
         * var azure = require('azure-storage');
         * var FileService = azure.createFileService();
         * var url = FileService.getUrl(shareName, directoryName, fileName, sasToken, true);
         */
-        getUrl(share: string, directory: string, file?: string, sasToken?: string, primary?: boolean): string;
+        getUrl(share: string, directory: string, file?: string, sasToken?: string, primary?: boolean, shareSnapshotId?: string): string;
 
         /**
         * Returns all user-defined metadata, standard HTTP properties, and system properties for the file.
@@ -6600,6 +6617,7 @@ declare module azurestorage {
         * @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
         * @param {Object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]                      Specifies the location mode used to decide which location the request should be sent to.
         *                                                                         Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]               The server timeout interval, in milliseconds, to use for the request.
@@ -6614,7 +6632,7 @@ declare module azurestorage {
         *                                                                         information about the file.
         *                                                                         `response` will contain information related to this operation.
         */
-        getFileProperties(share: string, directory: string, file: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.FileResult>): void;
+        getFileProperties(share: string, directory: string, file: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.FileResult>): void;
         getFileProperties(share: string, directory: string, file: string, callback: ErrorOrResult<FileService.FileResult>): void;
 
         /**
@@ -6625,6 +6643,8 @@ declare module azurestorage {
         * @param {string}             share                                       The share name.
         * @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
+        * @param {object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]                      Specifies the location mode used to decide which location the request should be sent to.
         *                                                                         Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]               The server timeout interval, in milliseconds, to use for the request.
@@ -6639,7 +6659,7 @@ declare module azurestorage {
         *                                                                         information about the file.
         *                                                                         `response` will contain information related to this operation.
         */
-        getFileMetadata(share: string, directory: string, file: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.FileResult>): void;
+        getFileMetadata(share: string, directory: string, file: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.FileResult>): void;
         getFileMetadata(share: string, directory: string, file: string, callback: ErrorOrResult<FileService.FileResult>): void;
 
         /**
@@ -6735,6 +6755,7 @@ declare module azurestorage {
         * @param {string}             directory                               The directory name. Use '' to refer to the base directory.
         * @param {string}             file                                    The file name. File names may not start or end with the delimiter '/'.
         * @param {Object}             [options]                               The request options.
+        * @param {string}             [options.shareSnapshotId]               The snapshot identifier of the share.
         * @param {LocationMode}       [options.locationMode]                  Specifies the location mode used to decide which location the request should be sent to.
         *                                                                     Please see StorageUtilities.LocationMode for the possible values.
         * @param {int}                [options.timeoutIntervalInMs]           The server timeout interval, in milliseconds, to use for the request.
@@ -6749,7 +6770,7 @@ declare module azurestorage {
         *                                                                     be true if the file exists, or false if the file does not exist.
         *                                                                     `response` will contain information related to this operation.
         */
-        doesFileExist(share: string, directory: string, file: string, options: common.RequestOptions, callback: ErrorOrResult<FileService.FileResult>): void;
+        doesFileExist(share: string, directory: string, file: string, options: FileService.FileServiceOptions, callback: ErrorOrResult<FileService.FileResult>): void;
         doesFileExist(share: string, directory: string, file: string, callback: ErrorOrResult<FileService.FileResult>): void;
 
         /**
@@ -6843,6 +6864,7 @@ declare module azurestorage {
         * @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
         * @param {Object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {int}                [options.rangeStart]                        The range start.
         * @param {int}                [options.rangeEnd]                          The range end.
         * @param {boolean}            [options.disableContentMD5Validation]       When set to true, MD5 validation will be disabled when downloading files.
@@ -6872,6 +6894,7 @@ declare module azurestorage {
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
         * @param {string}             localFileName                               The local path to the file to be downloaded.
         * @param {Object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {string}             [options.rangeStart]                        Return only the bytes of the file in the specified range.
         * @param {string}             [options.rangeEnd]                          Return only the bytes of the file in the specified range.
         * @param {boolean}            [options.useTransactionalMD5]               When set to true, Calculate and send/validate content MD5 for transactions.
@@ -6907,6 +6930,7 @@ declare module azurestorage {
         * @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
         * @param {Object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {string}             [options.rangeStart]                        Return only the bytes of the file in the specified range.
         * @param {string}             [options.rangeEnd]                          Return only the bytes of the file in the specified range.
         * @param {boolean}            [options.useTransactionalMD5]               When set to true, Calculate and send/validate content MD5 for transactions.
@@ -6942,6 +6966,7 @@ declare module azurestorage {
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
         * @param {Stream}             writeStream                                 The write stream.
         * @param {Object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {string}             [options.rangeStart]                        Return only the bytes of the file in the specified range.
         * @param {string}             [options.rangeEnd]                          Return only the bytes of the file in the specified range.
         * @param {boolean}            [options.useTransactionalMD5]               When set to true, Calculate and send/validate content MD5 for transactions.
@@ -6979,6 +7004,7 @@ declare module azurestorage {
         * @param {string}             directory                                   The directory name. Use '' to refer to the base directory.
         * @param {string}             file                                        The file name. File names may not start or end with the delimiter '/'.
         * @param {Object}             [options]                                   The request options.
+        * @param {string}             [options.shareSnapshotId]                   The snapshot identifier of the share.
         * @param {int}                [options.rangeStart]                        The range start.
         * @param {int}                [options.rangeEnd]                          The range end.
         * @param {LocationMode}       [options.locationMode]                      Specifies the location mode used to decide which location the request should be sent to.
@@ -7360,6 +7386,7 @@ declare module azurestorage {
 
         export interface ShareResult {
           name: string;
+          snapshot?: string;
           etag: string;
           lastModified: string;
           metadata?: { [key: string]: string; };
@@ -7423,12 +7450,20 @@ declare module azurestorage {
           quota: number
         }
 
-        export interface ListRangeRequestOptions extends common.RequestOptions {
+        export interface FileServiceOptions extends common.RequestOptions {
+          shareSnapshotId?: string;
+        }
+
+        export interface DeleteShareOptions extends FileServiceOptions {
+          deleteSnapshots?: string;
+        }
+
+        export interface ListRangeRequestOptions extends FileServiceOptions {
           rangeStart?: number;
           rangeEnd?: number;
         }
 
-        export interface GetFileRequestOptions extends common.RequestOptions {
+        export interface GetFileRequestOptions extends FileServiceOptions {
           parallelOperationThreadCount?: number;
           rangeStart?: number;
           rangeEnd?: number;
@@ -7436,7 +7471,12 @@ declare module azurestorage {
           disableContentMD5Validation?: boolean;
         }
 
-        export interface ListRequestOptions extends common.RequestOptions {
+        export interface ListRequestOptions extends FileServiceOptions {
+          maxResults?: number;
+          include?: string;
+        }
+
+        export interface ListShareRequestOptions extends common.RequestOptions {
           maxResults?: number;
           include?: string;
         }
@@ -7543,6 +7583,10 @@ declare module azurestorage {
           OFF: string;
           SHARE: string;
           FILE: string;
+        };
+
+        var ShareSnapshotDeleteOptions: {
+          SHARE_AND_SNAPSHOTS: string
         };
       }
     }