+++ ChangeLog.md
@@ -1,6 +1,10 @@
 Note: This is an Azure Storage only package. The all up Azure node sdk still has the old storage bits in there. In a future release, those storage bits will be removed and an npm dependency to this storage node sdk will 
 be taken. This is a GA release and the changes described below indicate the changes from the Azure node SDK 0.9.8 available here - https://github.com/Azure/azure-sdk-for-node.
 
+FILE
+
+* Added support for file share snapshot.
+
 2017.09 Version 2.5.0
 
 ALL